==> Tripadvisor/3. TF-IDF count.ipynb <==
import pandas as pd
import numpy as np

By_sentence_pos=pd.read_csv('By_sentence_pos.csv')

By_sentence_pos['pos']=By_sentence_pos['pos'].apply(lambda x:eval(x))
By_sentence_pos['token_ckip']=By_sentence_pos['pos'].apply(lambda x:[i[0].upper().strip() for i in x ])

#### 計算TF-IDF

tk =[]
words = []
for review in By_sentence_pos['token_ckip']:
    words = []
    for i in review:
        words.append(i)
    tk.append(words)

token_str=[]
for i in tk:
    text = ' '.join(i)
    token_str.append(text)

content = " ".join(token_str)
corpus =[content]

import os  
import sys  
from sklearn import feature_extraction  
from sklearn.feature_extraction.text import TfidfTransformer  
from sklearn.feature_extraction.text import CountVectorizer 

vectorizer=CountVectorizer()                              #该类会将文本中的词语转换为词频矩阵，矩阵元素a[i][j] 表示j词在i类文本下的词频  
transformer=TfidfTransformer()                            #该类会统计每个词语的tf-idf权值  
tfidf=transformer.fit_transform(vectorizer.fit_transform(corpus))#第一个fit_transform是计算tf-idf，第二个fit_transform是将文本转为词频矩阵  
word=vectorizer.get_feature_names()                       #获取词袋模型中的所有词语  
weight=tfidf.toarray()                                    #将tf-idf矩阵抽取出来，元素a[i][j]表示j词在i类文本中的tf-idf权重  


TFIDF=[]
for i in range(len(weight)):                              #打印每类文本的tf-idf词语权重，第一个for遍历所有文本，第二个for便利某一类文本下的词语权重  
    for j in range(len(word)): 
        a=[word[j],weight[i][j]]
        TFIDF.append(a)

tfidf= pd.DataFrame(TFIDF,columns=['word', 'weight'])

all_word=[]
for i in range(len(tk)):
    for j in tk[i]:
        all_word.append(j)
        
word_dup= list(set(all_word))

freq=[]
for i in word_dup:
    qq=[i,all_word.count(i)]
    freq.append(qq)

frequency=pd.DataFrame(freq,columns=["word",'frequency'])

word_tfidf=frequency.merge(tfidf,how='left')

word_tfidf.to_excel('word_tfidf.xlsx')
